alterações para os appointments e apagar loja

+++ src/Modules/appointment/useCases/delete.ts
@@ -1,10 +1,18 @@
 import { prisma } from "../../../Prisma/client";
+import { getByAppointmentId } from "../../appointment_client/useCases/getByAppointmentId";
 import { CreateAppointmentDTO } from "../dtos/CreateAppointmentDTO";
 
-export async function del( id :number ) {
-  // Verificar se o usuário já existe
+export async function del(id: number) {
+  const haveAppointments = await getByAppointmentId(id);
 
-  // Criar o usuário
+  if (haveAppointments && haveAppointments.data.length > 0) {
+    return {
+      sucess: false,
+      data: null,
+      message: "Não é possível excluir o horário, pois há agendamentos",
+    };
+  }
+
   const data = await prisma.appointment.delete({
     where: { id: id },
   });
